+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changes
 Unreleased
 ----------
 
+- Add support for PoP key (cnf claim). `#50 <https://github.com/dajiaji/python-cwt/pull/50>`__
+- Add to_dict to COSEKey. `#50 <https://github.com/dajiaji/python-cwt/pull/50>`__
+- Add crv property to COSEKey. `#50 <https://github.com/dajiaji/python-cwt/pull/50>`__
+- Add key property to COSEKey. `#50 <https://github.com/dajiaji/python-cwt/pull/50>`__
 - Add support for RSASSA-PSS. `#49 <https://github.com/dajiaji/python-cwt/pull/49>`__
 - Add support for RSASSA-PKCS1-v1_5. `#48 <https://github.com/dajiaji/python-cwt/pull/48>`__